refactor(tensor): split core.py
