+++ quanto/tensor/__init__.py
@@ -1,3 +1,5 @@
 from .core import *
 from .packed import *
+from .qbitstensor import *
+from .qtensor import *
 from .qtype import *
